Handle errors after rebase Moved login to separate auth module Moved config to common module Refactored class names
