+++ homebase-auth/src/commonMain/kotlin/id/homebase/auth/login/LoginScreen.kt
@@ -1,21 +1,40 @@
-package id.homebase.chat.login
+package id.homebase.auth.login
 
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Button
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.Lifecycle
@@ -105,14 +124,11 @@ fun LoginScreen(
                 uiState.errorMessage != null -> LoginError(
                     message = uiState.errorMessage ?: "",
                     homebaseId = uiState.homebaseId,
-                    onHomebaseIdChange = { viewModel.onAction(LoginUiAction.HomebaseIdChanged(it)) },
-                    onRetryClick = { viewModel.onAction(LoginUiAction.RetryClicked) }
+                    onRetryClick = { viewModel.onAction(LoginUiAction.RetryClicked(it)) }
                 )
 
                 else -> LoginForm(
-                    homebaseId = uiState.homebaseId,
-                    onHomebaseIdChange = { viewModel.onAction(LoginUiAction.HomebaseIdChanged(it)) },
-                    onLoginClick = { viewModel.onAction(LoginUiAction.LoginClicked) }
+                    onLoginClick = { viewModel.onAction(LoginUiAction.LoginClicked(it)) }
                 )
             }
         }
@@ -147,11 +163,10 @@ private fun LoginSuccess() {
 
 @Composable
 private fun LoginForm(
-    homebaseId: String,
-    onHomebaseIdChange: (String) -> Unit,
-    onLoginClick: () -> Unit,
+    onLoginClick: (homebaseId: String) -> Unit,
 ) {
     val focusRequester = remember { FocusRequester() }
+    var homebaseId by remember { mutableStateOf("") }
 
     LaunchedEffect(Unit) {
         focusRequester.requestFocus()
@@ -163,15 +178,17 @@ private fun LoginForm(
     ) {
         HomebaseIdField(
             value = homebaseId,
-            onValueChange = onHomebaseIdChange,
+            onValueChange = {
+                homebaseId = it
+            },
             focusRequester = focusRequester,
-            onDone = onLoginClick
+            onDone = { onLoginClick(homebaseId) }
         )
 
         Spacer(modifier = Modifier.height(24.dp))
 
         Button(
-            onClick = onLoginClick,
+            onClick = { onLoginClick(homebaseId) },
             modifier = Modifier.fillMaxWidth()
         ) {
             Text("Sign In")
@@ -183,10 +200,10 @@ private fun LoginForm(
 private fun LoginError(
     message: String,
     homebaseId: String,
-    onHomebaseIdChange: (String) -> Unit,
-    onRetryClick: () -> Unit,
+    onRetryClick: (homebaseId: String) -> Unit,
 ) {
     val focusRequester = remember { FocusRequester() }
+    var homebaseId by remember { mutableStateOf(homebaseId) }
 
     LaunchedEffect(Unit) {
         focusRequester.requestFocus()
@@ -207,15 +224,17 @@ private fun LoginError(
 
         HomebaseIdField(
             value = homebaseId,
-            onValueChange = onHomebaseIdChange,
+            onValueChange = {
+                homebaseId = it
+            },
             focusRequester = focusRequester,
-            onDone = onRetryClick
+            onDone = { onRetryClick(homebaseId) }
         )
 
         Spacer(modifier = Modifier.height(24.dp))
 
         Button(
-            onClick = onRetryClick,
+            onClick = { onRetryClick(homebaseId) },
             modifier = Modifier.fillMaxWidth()
         ) {
             Text("Try Again")
@@ -233,11 +252,8 @@ private fun HomebaseIdField(
     onDone: () -> Unit,
 ) {
     OutlinedTextField(
-        value = TextFieldValue(
-            text = value,
-            selection = TextRange(value.length)
-        ),
-        onValueChange = { onValueChange(it.text) },
+        value = value,
+        onValueChange = { onValueChange(it) },
         modifier = Modifier
             .fillMaxWidth()
             .focusRequester(focusRequester),
