+++ homebase-auth/src/commonMain/kotlin/id/homebase/auth/login/LoginUiAction.kt
@@ -1,0 +1,14 @@
+package id.homebase.auth.login
+
+sealed interface LoginUiAction {
+
+    data class LoginClicked(val homebaseId: String) : LoginUiAction
+
+    data class RetryClicked(val homebaseId: String)  : LoginUiAction
+
+    /**
+     * App returned to foreground (e.g. browser auth cancelled or completed)
+     */
+    data object AppResumed : LoginUiAction
+}
+
