+++ homebase-chat/src/commonMain/kotlin/id/homebase/chat/ConversationListUiEvent.kt
@@ -1,0 +1,5 @@
+package id.homebase.chat
+
+sealed interface ConversationListUiEvent {
+    data object NavigateBack : ConversationListUiEvent
+}
