+++ homebase-common/src/commonMain/kotlin/id/homebase/core/util/ScrollPosition.kt
@@ -1,0 +1,6 @@
+package id.homebase.core.util
+
+data class ScrollPosition(
+    val firstVisibleItemIndex: Int = 0,
+    val firstVisibleItemScrollOffset: Int = 0
+)
